pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-digest"
version = "0.2.0"
description = "Japanese language shadowing tool with email digest"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.07.0",
    "openai>=1.40.0",
    "resend>=2.3.0",
    "python-dotenv>=1.0.0",
    "openai-whisper",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.12",
]

[project.scripts]
daily-digest = "src.cli:main"
